refactor(v3): dedup logger normalize, share LogFilePoller, trim comments
Three maintainability actions from the V3 audit plan (#2, #4 of 6).

1. LogRouter.Normalize(entry) consolidates the LogEntry normalization
   (path normalization + host-field stamping from Environment) that
   was copy-pasted between JsonDailyLogger.Append and
   XmlDailyLogger.Append (~10 lines × 2). Both loggers now call the
   helper. XmlDailyLogger's private ToNormalizedPath wrapper becomes
   unused → removed.

2. LogFilePoller.WaitForLinesAsync extracted to a shared test helper
   in tests/LogCentralizer.Tests/. The polling loop (~25 lines) was
   duplicated verbatim between LogCentralizerTests and
   LogCentralizerE2ETests with only a poll-delay constant differing.
   Both call sites updated; private copies removed.

3. CryptoSoftAdapter.GlobalMutexName promoted from private const to
   internal const. The test that re-declared the same literal now
   references it directly via InternalsVisibleTo — a future rename
   of the constant correctly breaks the test instead of silently
   pinning the old name.

4. Comment elagage (~120 lines removed across LogCentralizer/Program.cs
   header, JsonDailyLogger.FlushBatch belt-and-braces blocks,
   CryptoSoftAdapter ctor/Encrypt). Targets: refs to merged PRs that
   rot ("issue #155", "Pre-fix"), narration of past refactors
   ("previous version did X"), and explanations of WHAT the code does
   (well-named identifiers already cover that). Kept all WHY-non-obvious
   comments (hidden invariants, OS-specific quirks, contract notes).

Net diff: -154 lines across 8 files + 1 new file (LogFilePoller, 38 lines).

Verified:
- dotnet build EasySave.sln (10 projects, 0 errors)
- dotnet format --verify-no-changes (clean)
- All non-flaky tests pass: V1 125/125, V2 154/160 (4 macOS-only
  SelfSignedCertProviderTests pre-existing, 2 skipped), LogCentralizer
  7 in-process pass.
