+++ src/EasyLog/LogRouter.cs
@@ -2,12 +2,13 @@ namespace EasyLog;
 
 /// <summary>
 /// Shared helpers used by <see cref="JsonDailyLogger"/> and
-/// <see cref="XmlDailyLogger"/> to apply the V3 <see cref="LogMode"/>
-/// routing rules in a single place. Pulling these out of the loggers
-/// keeps the per-format Append paths free of duplicated bookkeeping
-/// (CLAUDE.md zero-duplication grading criterion) and ensures both
-/// formats follow the exact same fall-back semantics when a misconfigured
-/// host wires a centralized mode with a null shipper.
+/// <see cref="XmlDailyLogger"/> for the V3 <see cref="LogMode"/> routing
+/// rules and for the canonical normalization of an incoming
+/// <see cref="LogEntry"/>. Centralizing these here keeps the per-format
+/// Append paths free of duplicated bookkeeping (CLAUDE.md zero-duplication
+/// criterion) and guarantees both formats follow the exact same fall-back
+/// semantics when a misconfigured host wires a centralized mode with a
+/// null shipper.
 /// </summary>
 internal static class LogRouter
 {
@@ -27,4 +28,27 @@ internal static class LogRouter
     /// <summary>True when the entry should be persisted to the local daily file.</summary>
     public static bool ShouldWriteLocal(LogMode mode) =>
         mode is LogMode.Local or LogMode.Both;
+
+    /// <summary>
+    /// Returns a normalized copy of <paramref name="entry"/> ready for
+    /// persistence: source / target paths run through
+    /// <see cref="LogPathHelper.ToNormalizedPath"/>, and host fields are
+    /// stamped from <see cref="Environment"/> when the caller left them
+    /// null. Caller-provided values are preserved untouched so a central
+    /// collector relaying entries from remote hosts never overwrites the
+    /// original sender's identity.
+    /// </summary>
+    public static LogEntry Normalize(LogEntry entry) => new()
+    {
+        Timestamp = entry.Timestamp,
+        JobName = entry.JobName,
+        SourceFile = LogPathHelper.ToNormalizedPath(entry.SourceFile),
+        TargetFile = LogPathHelper.ToNormalizedPath(entry.TargetFile),
+        FileSize = entry.FileSize,
+        FileTransferTimeMs = entry.FileTransferTimeMs,
+        EncryptionTimeMs = entry.EncryptionTimeMs,
+        EventType = entry.EventType,
+        MachineName = entry.MachineName ?? Environment.MachineName,
+        UserName = entry.UserName ?? Environment.UserName,
+    };
 }
